Manage file processing in memory

+++ src/main/java/com/example/cleanline/service/FileCleaner.java
@@ -8,7 +8,9 @@ import java.nio.file.Paths;
 import java.text.SimpleDateFormat;
 import java.util.Date;
 import java.util.HashSet;
+import java.util.LinkedHashSet;
 import java.util.Set;
+import java.util.stream.Collectors;
 
 import com.example.cleanline.utils.FileUtils;
 
@@ -17,10 +19,31 @@ public class FileCleaner {
     public static final Charset UTF8 = Charset.forName("UTF-8");
     FileUtils fileUtils = new FileUtils();
 
+    public String removeDuplicateLines(String unprocessedFileContent) {
+        System.out.println("Removing deduplicate lines");
+        String dedup = unprocessedFileContent.lines()
+        .collect(Collectors.collectingAndThen(Collectors.toCollection(LinkedHashSet::new),
+        set -> String.join(System.lineSeparator(), set)));
+        System.out.println("Lines deduplicated");
+
+        return dedup;
+    }
+
+    public String removeEmptyLines(String unprocessedFileContent) {
+        System.out.println("Removing empty lines");
+        String nonEmpty = unprocessedFileContent.lines()
+        .filter(s -> !s.isEmpty())
+        .collect(Collectors.joining(System.lineSeparator()));
+        System.out.println("Empty lines removed");
+
+        return nonEmpty;
+    }
+
     /*
      * Remove empty lines from a text file
      */
-    public void removeEmptyLines(File inputFile) throws IOException {
+    @Deprecated
+    public void removeEmptyLinesAndWriteToFile(File inputFile) throws IOException {
         File tempFile = new File(inputFile.getName()+".tmp");
         String originalPath = inputFile.getAbsolutePath();
         try(BufferedReader bufferedReader = new BufferedReader(new InputStreamReader(new FileInputStream(inputFile), UTF8));
@@ -43,27 +66,33 @@ public class FileCleaner {
     /*
      * Remove duplicate lines from a text file
      */
-    public void removeDuplicateLines(File inputFile) throws IOException {
-        File tempFile = new File(inputFile.getName()+".tmp");
-        String originalPath = inputFile.getAbsolutePath();
-        try(BufferedReader bufferedReader = new BufferedReader(new InputStreamReader(new FileInputStream(inputFile), UTF8));
-            PrintWriter printWriter = new PrintWriter(new OutputStreamWriter(new FileOutputStream(tempFile), UTF8))) {
-                System.out.println("Removing duplicate lines");
-                String line = null; 
-                Set<String> lines = new HashSet<String>();
-                while ((line = bufferedReader.readLine()) != null) {
-                    lines.add(line);
+    @Deprecated
+    public void removeDuplicateLinesAndWriteToFile(File inputFile) throws IOException {
+        if (inputFile != null) {
+            File tempFile = new File(inputFile.getName()+".tmp");
+            String originalPath = inputFile.getAbsolutePath();
+            try(BufferedReader bufferedReader = new BufferedReader(new InputStreamReader(new FileInputStream(inputFile), UTF8));
+                PrintWriter printWriter = new PrintWriter(new OutputStreamWriter(new FileOutputStream(tempFile), UTF8))) {
+                    System.out.println("Removing duplicate lines");
+                    String line = null; 
+                    Set<String> lines = new HashSet<String>();
+                    while ((line = bufferedReader.readLine()) != null) {
+                        lines.add(line);
+                    }
+                    for (String nonEmptyLine : lines) {
+                        printWriter.println(nonEmptyLine);
+                    }
+                    System.out.println("Duplicate lines removed");
+                } catch(IOException e) {
+                    e.printStackTrace();
                 }
-                for (String nonEmptyLine : lines) {
-                    printWriter.println(nonEmptyLine);
-                }
-                System.out.println("Duplicate lines removed");
-            } catch(IOException e) {
-                e.printStackTrace();
-            }
-            fileUtils.deleteFile(inputFile); 
-            fileUtils.moveFile(tempFile.toPath(), Paths.get(originalPath));
-    }
+                fileUtils.deleteFile(inputFile); 
+                fileUtils.moveFile(tempFile.toPath(), Paths.get(originalPath));
+        }
+        else {
+            System.out.println("Input file is null");
+        }
+     }
 
     /*
      * Backup file
@@ -72,24 +101,29 @@ public class FileCleaner {
 
         boolean isBackedUp = false; 
 
-        long timestamp = System.currentTimeMillis();
-        Date date = new Date(timestamp);
+        if (selectedFile != null) {
+            long timestamp = System.currentTimeMillis();
+            Date date = new Date(timestamp);
 
-        SimpleDateFormat simpleDateFormat = new SimpleDateFormat("yyyy-MM-dd_HH:mm:ss");
-        String formattedDate = simpleDateFormat.format(date);
+            SimpleDateFormat simpleDateFormat = new SimpleDateFormat("yyyy-MM-dd_HH:mm:ss");
+            String formattedDate = simpleDateFormat.format(date);
 
-        Path sourceFile = Paths.get(selectedFile.getAbsolutePath());
-        Path backupFile = Paths.get(selectedFile.getAbsolutePath() + "_" + formattedDate);
+            Path sourceFile = Paths.get(selectedFile.getAbsolutePath());
+            Path backupFile = Paths.get(selectedFile.getAbsolutePath() + "_" + formattedDate);
 
-         try {
-            Files.copy(sourceFile, backupFile);
-            System.out.println("Backup file successfully created");
-            isBackedUp = true; 
-        } catch (IOException e) {
-            System.out.println("Error while creating backup file: " + e.getMessage());
-            isBackedUp = false; 
+            try {
+                Files.copy(sourceFile, backupFile);
+                System.out.println("Backup file successfully created");
+                isBackedUp = true; 
+            } catch (IOException e) {
+                System.out.println("Error while creating backup file: " + e.getMessage());
+                isBackedUp = false; 
+            }
         }
-        
+        else {
+            System.out.println("Selected file is null");
+        }
+
         return isBackedUp;
     }
